// client/src/types.ts
export interface User {
  username: string;
  uid?: string;
}
// ...
export interface Comment {
  text: string;
  commentBy: string;
  commentDateTime: Date;
}
// ...
export interface RootsState {
  user: User | null;
  dataStorageReducer: {
    datastorages: UserSettings;
  };
}

/**
 * Type for user UI settings.
 */
export interface UserSettings {
  globalMode: string | null;
  headerMode: string | null;
  sidebarMode: string | null;
  textMode: string | null;
  buttonMode: string | null;
}